cleanup api a bit
