+++ dask/dataframe/io/dataset/__init__.py
@@ -1,0 +1,1 @@
+from dask.dataframe.io.dataset.core import from_dataset
